requirements.txt
fastapi
uvicorn[standard]
motor
httpx
pydantic
pytest
pytest-asyncio
python-dotenv